Feature: utils reorganize
